feat(Initial Data): Route Stop Map
Have create a function which returns a hash map from stops to a list of routes and directions. Test
was created and firest used with raw data from LTC. I grabbed a sample of that data in sample.csv.
Then I jsut copied to douplicate the stops by changing the first column value to a different route.
This allowed me to test the building of the list