feat(Data): BusStopMap
Added in the route name to the list of route objects. Which is the value to the stopID key. I have
done this by bringing in a array of object which have the number and name for the routes. I user
R.find with the predicate R.propEq('number', routeNumber) to search the array and return the object.
Then grabbing the name property I pass it to a capitialize funciton