feat(InitialData): OMG
I got a script written to scrap the stop data! All I have to type it npm run scrap -- [routeID]
[direction]. Then x-ray will go to that given route and direciton to grab all the stops. Then writes
to a file with the routeID-direction.json signature. I would like to find a way to grab the list of
routes.json and use it to excute child spawns to automatically run this scrapping script.