feat(initialData): reducer and bundling json
I have created a script to bundle the .json files in the Stops folder into one file call
allStops.js. This file will then be used by createBusStopsMap in core.js to generate the stopIDMap.
I have created the reducer to handle this along with the constant and action funciton. There are
tests created to test the action and the reducer for generating a stopIDMap. However it comes from
the sample.json file which is a random copy&paste from allStops.json.