feat(Data): API and redux-form
Installed redux-form and have the very basics set up. Still need to deal with the clearing and
submitting fucntions. I have the state in an immutable Map using redux-immutable. This required me
to look into redux-form's FAQs to the one on connecting to an immutable data state tree. I have the
stops rendering to the screen. Need to add the filtering functionality next by using the users
input. This can be done with the onSubmit function I pass to handleSubmit. I have the initial json
data now coming from the express server instead of just importing the file. The change is reflected
in the server.js file.

close #3