feat(Storage): FilteredMap
I have created a new part of the store, called filteredMap. This will be apart of the new storage
reducer. Also made the required constants and actionCreators. Redux-localstorage is written wirte
now in the configureStore.js file. This will all have to be refactored in a dev and prod version.
However whenever there is a change to the store now the filteredMap will get serialized and stored.
Then it is merged with that part of the state upon return of the user. At this moment the list being
rendered is still based off the stopIDMap. Meaning the user does not see the filtered list at this
point. Next set is to find a way to render the filtered list if it exists in in the constructor.
However the full list needs to be rendered and filtered every other time. Just not sure how to
handle this yet