style(List): styling
I have done mostly styling in this commit. I did change however the dynamicly updating list upon
user input to be a search field that has to be submited to run. On mobile the keyboard would block
out the list anyway. Therefore I felt this was a waste of processing power. Plus I notcied a delay
when the user would be typing a stopID and then go to clear the input field. A stutter. So I felt it
best to go with a normal search. I will need to hide the button next, no worries.