feat(UX): Fetching data
I have improved the fetching of the data I believe. By creating a reducer to handle the state of the
fetching. Done by dispatching an action to set a flag to true, which cases a spinner to show. When
the data come back another action is dispatched to set fetching to false and success to true. I
added in a setTimout delay for when there is no list in localStorage. This is to simulate getting
the massive amount of stops.