perf(Data): Initial Data file
Instead of sending the long list of stop from allStops.json and then convert once the user loads the
page with createStopsMap form core.js. I made a script in utils called convertStopsFile.js. This
script will take in the allStop.json file run the createStopsMap function and output the state in
the shape I want. With a list of object stops keys and a array of route objects. I changed in the
initialData reducer to just using fromJS on the returned data now that it is in the desired format
coming from the server. Should help with performance.