chore(Server): split modes
I have split the server.js file into a dev and prod version so that it is less confusing. This is in
preperation for deploying on heroku. I had to stop the server from always building with webpack on
the fly. There are some updates to the README file as well to refeclt the direction the app has
taken. I have modifyied the travis file aswell. I should do a complete overhaul to include
semantic-versioning.