chore(heroku): deployment
I am working on getting the deployment to heroku.

refactor(Data): Initial data from mongo

When deploying to heroku I realized I was still trying to load the initial data form the json files
that are not included in the git push. I have change the api route for the server to fetch all the
stops form the mongoDB now and send to the client.